Add Questions class
Separates the difference between levels and questions. This means it's
now possible to have multiple questions in a single level.

There is a WriteCodeQuestion that evaluates the code the user submits
and evaluates it with a test suite. This builds a python script based
on strings in the code, which means the generated python script might
get bugs in the future, from simply not having any way of being linted
or tested. However, this was the simplest way found of implementing the
code evaluation of the users' code, while also running a test suite,
since it doesn't require us to set up a service like Snekbox.

We are aware that users with significant python skill may be able to
bypass the test regex, but since the rewards for bypassing the filter is
only your own character progressing in the game, it is deemed neither a
security risk nor a risk to other players of the game.

Since the code is run in a sandboxed environment with Tio.run,
constructing a code snippet that will run, with user input simply
inserted into the code, poses no security risk. The worst thing that can
happen is that the user completes the game quickly, which only ruins the
fun for themselves.