Add title screen to play command
Renames /map to /play as well, to serve as an entry point to the game.

The StoryView that's implemented here will be possible to use for
success screens, fail screens and future story segments in the game.
There is also a skippable option, to decide if the user should be able
to skip the story, or if the story must be completed. This can be
useful, for example, when re-playing the game a second time and don't
want all the story again.