Remove valueerror when loading levels multiple times
When the bot is temporarily disconnected, but the session still remains, when the bot is reconnected  will run a second time. Since load levels was done inside on_ready, it would load the levels a second time. Since the levels have already been added, it raised a ValueError. This commit changes where the levels are loaded, to not have this issue